condense all env var util functions to a single file
