--- src/utils/getApiUrl.ts
@@ -1,1 +1,0 @@
-export const getApiUrl = () => process.env.REACT_APP_API_URL || "";
