--- src/utils/getUiUrl.ts
@@ -1,1 +1,0 @@
-export const getUiUrl = () => process.env.REACT_APP_UI_URL || "";
