--- src/utils/isDevelopment.ts
@@ -1,1 +1,0 @@
-export const isDevelopment = () => process.env.NODE_ENV === "development";
