--- src/utils/isTest.ts
@@ -1,1 +1,0 @@
-export const isTest = () => process.env.NODE_ENV === "test";
